KP-7378 Fix wonky image placement on some slides
When we have content outside a column div next to a column div, the
latter will not be located where a column normally would, but instead it
starts lower down the slide. This is not desired. To match the visuals
in the pdf slides, I added the second, implied, column div into the
source markdown. Now `convert.py` from slidefactory git installation
produces the expected results.