KP-7378 Only deploy from main
This way we won't overwrite the production website with unfinished
content that has not been reviewed yet. Left the extra rows as comments,
because enabling them allows one to provision the full website on their
own fork (and thus in myuser.github.io/Kielipankki instead of on the
official course material site).