KP-7421 Provide the front page url to the tutorial converting script
This way it can be used when adding the navigation link. The URL is
fetched from the GitHub API to ensure that the home page link also works
in forked repositories. This should avoid missing bugs because one
accidentally clicked a home page link that took them to the production
(where a feature isn't broken − yet).